Moved summarizers into core
