+++ griptape/core/summarizers/__init__.py
@@ -1,0 +1,7 @@
+from griptape.core.summarizers.summarizer import Summarizer
+from griptape.core.summarizers.prompt_driver_summarizer import PromptDriverSummarizer
+
+__all__ = [
+    "Summarizer",
+    "PromptDriverSummarizer"
+]
