--- griptape/summarizers/__init__.py
@@ -1,7 +1,0 @@
-from griptape.summarizers.summarizer import Summarizer
-from griptape.summarizers.prompt_driver_summarizer import PromptDriverSummarizer
-
-__all__ = [
-    "Summarizer",
-    "PromptDriverSummarizer"
-]
